requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
huggingface-hub==0.16.4
torch==2.1.0
transformers==4.34.0
tokenizers==0.14.0
pydantic==2.4.2
python-dotenv==1.1.0
httpx==0.24.1
aiosqlite==0.19.0
schedule==1.2.0 